refactor: decouble event repository from event data source
